+++ zkml/src/lib.rs
@@ -21,7 +21,7 @@ pub mod model;
 mod onnx_parse;
 pub mod padding;
 mod parser;
-pub use onnx_parse::{FloatOnnxLoader, ModelType};
+pub use parser::{FloatOnnxLoader, ModelType};
 pub mod tensor;
 pub use tensor::Tensor;
 #[cfg(test)]
@@ -174,7 +174,7 @@ mod test {
     use crate::{
         FloatOnnxLoader, default_transcript,
         iop::{Context, prover::Prover, verifier::verify},
-        onnx_parse::ModelType,
+        parser::ModelType,
         tensor::Tensor,
         to_bit_sequence_le,
     };
